Refactor task classes into Tasks package
