+++ src/main/java/Rakan/Tasks/Deadline.java
@@ -1,9 +1,9 @@
-package Rakan;
+package Rakan.Tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task{
+public class Deadline extends Task {
 
     protected LocalDateTime by;
