+++ src/main/java/Rakan/Tasks/Event.java
@@ -1,9 +1,9 @@
-package Rakan;
+package Rakan.Tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Event extends Task {
 
     protected LocalDateTime from;
     protected LocalDateTime to;
