+++ src/main/java/Rakan/Tasks/ToDo.java
@@ -1,6 +1,6 @@
-package Rakan;
+package Rakan.Tasks;
 
-public class ToDo extends Task{
+public class ToDo extends Task {
     public ToDo(String description) {
         super(description);
     }
